+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Change climate target temperature on floor plan.
+
 ## [2026.7.2] - 2026-07-23
 
 ### Fixed